Refactor root path response and add new route for dynamic username and id
In index.js, the code has been refactored to change the response for the root path from "Hello World!" to "This is a root path". Additionally, a new route has been added that accepts dynamic parameters for username and id, and responds with a customized message. The console will log "request received" when this route is accessed.